Tried to stop processes when configs change
Seems like that the run directory is not used by the hadoop script.
Therefore when executing 'stop DataNode' a reply no datanode to stop
comes. But a datanode runs. Therefore when a config file is synched
wrong and then the datanode is started there is now way to recover from
that. Not even a kill works, because of the limitations of the hduser I
guess.